refactor(blorktools): animation and rig directories
