+++ packages/blorktools/src/asset_debugger/html-editor-modal/html-editor-modal.html
@@ -1,11 +1,6 @@
-<link rel="stylesheet" href="../../css/html-editor-modal.css">
-<style>
-    /* Hide animation dropdown when long-exposure-mode class is present */
-    #editor-dropdowns-container.long-exposure-mode .dropdown-hint:nth-child(5),
-    #editor-dropdowns-container.long-exposure-mode #html-animation-type {
-        display: none;
-    }
-</style>
+<!-- Link to external stylesheet -->
+<link rel="stylesheet" href="/asset_debugger/html-editor-modal/html-editor-modal.css">
+
 <!-- Embedded HTML Editor Modal -->
 <div id="html-editor-modal" class="modal-overlay">
     <div class="modal-container">
