+++ packages/blorktools/src/asset_debugger/rig/rig-panel.html
