+++ packages/blorktools/src/asset_debugger/ui/pages/asset_debugger.html
@@ -6,7 +6,6 @@
     <title>Asset Debugger - Blorktools</title>
     <link rel="stylesheet" href="../../css/styles.css">
     <link rel="stylesheet" href="../../css/loading-splash.css">
-    <link rel="stylesheet" href="../../css/html-editor-modal.css">
     <script>
         // Get pinned state from localStorage before DOM is constructed
         function getPinnedState() {
