+++ packages/blorktools/src/asset_debugger/ui/scripts/settings-modal.js
@@ -2,7 +2,7 @@
  * Settings Modal UI Component for Asset Debugger
  */
 import { saveSettings, loadSettings, getDefaultSettings } from '../../data/localstorage-util.js';
-import { updateRigOptions } from '../../core/rig/rig-manager.js';
+import { updateRigOptions } from '../../rig/rig-manager.js';
 
 export class SettingsModal {
     constructor(settings = null) {
