--- packages/blorktools/src/asset_debugger/ui/styles/html-editor.css
@@ -1,99 +1,0 @@
-/* HTML Editor Styles */
-
-/* Error styling for textarea */
-.html-editor-textarea.has-errors {
-    border: 1px solid #dc3545 !important;
-    background-color: rgba(220, 53, 69, 0.05);
-}
-
-/* Error container */
-.html-editor-errors {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    max-height: 100px;
-    overflow-y: auto;
-    background-color: #f8d7da;
-    color: #721c24;
-    border-top: 1px solid #f5c6cb;
-    padding: 8px;
-    font-size: 12px;
-    z-index: 10;
-}
-
-.html-editor-errors ul {
-    margin: 0;
-    padding: 0 0 0 20px;
-}
-
-.html-editor-errors li {
-    cursor: pointer;
-    margin-bottom: 4px;
-}
-
-.html-editor-errors li:hover {
-    text-decoration: underline;
-}
-
-/* Status messages */
-.editor-status {
-    padding: 5px;
-    margin-top: 5px;
-    border-radius: 3px;
-    font-size: 12px;
-}
-
-.editor-status.success {
-    background-color: #d4edda;
-    color: #155724;
-}
-
-.editor-status.error {
-    background-color: #f8d7da;
-    color: #721c24;
-}
-
-.editor-status.info {
-    background-color: #d1ecf1;
-    color: #0c5460;
-}
-
-/* Editor container positioning */
-.editor-container {
-    position: relative;
-}
-
-/* Textarea styling */
-#html-editor-textarea {
-    width: 100%;
-    min-height: 300px;
-    font-family: monospace;
-    padding: 10px;
-    border: 1px solid #ced4da;
-    border-radius: 4px;
-    resize: vertical;
-}
-
-/* Preview container */
-#html-preview-container {
-    border: 1px solid #ced4da;
-    border-radius: 4px;
-    padding: 10px;
-    min-height: 300px;
-    background-color: white;
-}
-
-/* Format button styling */
-#html-editor-format {
-    background-color: #6c757d;
-    color: white;
-    border: none;
-    padding: 5px 10px;
-    border-radius: 3px;
-    cursor: pointer;
-}
-
-#html-editor-format:hover {
-    background-color: #5a6268;
-}
